# inspiracion/panda-etl/backend/app/docs/swagger.yaml
openapi: 3.0.0
info:
  title: MagicETL API
  description: |
    API para MagicETL - Plataforma modular de ETL con capacidades de IA.
    Esta API proporciona endpoints para gestionar proyectos, procesos ETL, transformaciones con IA y más.
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo MagicETL
    email: support@magicetl.com
servers:
  - url: /api/panda
    description: Servidor principal de MagicETL
...
paths:
...
  /v1/processes/{process_id}:
    get:
      summary: Obtiene un proceso por su ID
      tags: [Procesos]
      parameters:
        - name: process_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles del proceso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
        '404':
          description: Proceso no encontrado
...
  /v1/pipelines:
    get:
      summary: Obtiene todos los pipelines
      tags: [Pipelines]
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Lista de pipelines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pipeline'
    post:
      summary: Crea un nuevo pipeline
      tags: [Pipelines]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineCreate'
      responses:
        '201':
          description: Pipeline creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
  /v1/pipelines/{pipeline_id}:
    get:
      summary: Obtiene un pipeline por su ID
      tags: [Pipelines]
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalles del pipeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
        '404':
          description: Pipeline no encontrado
    delete:
      summary: Elimina un pipeline
      tags: [Pipelines]
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Pipeline eliminado
        '404':
          description: Pipeline no encontrado
  /v1/pipelines/{pipeline_id}/versions:
    get:
      summary: Obtiene todas las versiones de un pipeline
      tags: [Pipelines]
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de versiones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PipelineVersion'
    post:
      summary: Crea una nueva versión de un pipeline
      tags: [Pipelines]
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineVersionCreate'
      responses:
        '201':
          description: Versión creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineVersion'
  /v1/pipelines/{pipeline_id}/execute:
    post:
      summary: Ejecuta un pipeline
      tags: [Pipelines]
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
        - name: version_id
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineExecuteRequest'
      responses:
        '200':
          description: Resultado de la ejecución
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineExecutionResult'
components:
  schemas:
...
    Process:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, FAILED, STOPPED]
        project_id:
          type: integer
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        details:
          type: object
        message:
          type: string
        output:
          type: object
...
    Pipeline:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        project_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version_count:
          type: integer
        versions:
          type: array
          items:
            type: string
    PipelineCreate:
      type: object
      required:
        - name
        - project_id
      properties:
        name:
          type: string
        description:
          type: string
        project_id:
          type: integer
    PipelineVersion:
      type: object
      properties:
        version_id:
          type: string
        pipeline_id:
          type: string
        created_at:
          type: string
          format: date-time
        description:
          type: string
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
        transform_script:
          type: string
        config:
          type: object
    PipelineVersionCreate:
      type: object
      required:
        - nodes
        - edges
      properties:
        description:
          type: string
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
        transform_script:
          type: string
        config:
          type: object
    PipelineExecuteRequest:
      type: object
      properties:
        input_config:
          type: object
        output_config:
          type: object
    PipelineExecutionResult:
      type: object
      properties:
        pipeline_id:
          type: string
        version:
          type: string
        execution_time:
          type: number
        rows_processed:
          type: integer
        rows_output:
          type: integer
        columns_input:
          type: array
          items:
            type: string
        columns_output:
          type: array
          items:
            type: string
        executed_at:
          type: string
          format: date-time
        status:
          type: string